+++ local/resources/views/admin/sales/revisifakturupdate.blade.php
@@ -80,7 +80,7 @@
                                      <td>{{ $item['jumlahekor'] }}</td>
                                      <td>{{ $item['keterangan'] }}</td>
                                      <td>
-                                        {!! Form::open(['method' => 'DELETE', 'route'=>['admin.sales.detailrevisifaktur.destroy', $item['id'] ]]) !!}
+                                        {!! Form::open(['method' => 'DELETE', 'route'=>['admin.sales.detailrevisifaktur.destroy', $item['id']], 'onsubmit'=>'return confirm(\'Do you want to delete it\')']) !!}
                                         {!! Form::hidden('nojual', $jual->nojual) !!}
                                         {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                                         {!! Form::close() !!}
