+++ local/resources/views/owner/suppliers/index.blade.php
@@ -27,7 +27,7 @@
          <td>{{ $supplier->emailsupp }}</td>
          <td><a href="{{route('owner.suppliers.edit',$supplier->id)}}" class="btn btn-warning">Update</a></td>
          <td>
-         {!! Form::open(['method' => 'DELETE', 'route'=>['owner.suppliers.destroy', $supplier->id]]) !!}
+         {!! Form::open(['method' => 'DELETE', 'route'=>['owner.suppliers.destroy', $supplier->id], 'onsubmit'=>'return confirm(\'Do you want to delete it\')']) !!}
          {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
          {!! Form::close() !!}
          </td>
